'views.py'/purchase_dog_product()
This function gets the 'dog_product' associated with the kwarg provided.
if the product is found to be in stock the 'dog_product' quantity is
reduced by 1, a new Purchase is created, and the user is redirected to
'purchase_detail.html at the id of the purchase.

else if the product is not in stock, the user is prompeted the error
message "{dog_product.name} is out of stock" and redirected to
'purchase_detail.html at the id of the purchase.