views.py/urls.py change
urls.py:
path for the purchase detail was changed to the use the new class
from views.py

views.py:
NEW CLASS: PurchaseDetailView
This class will be replacing the previous function that was in its place
PurchaseDetailView will do the exact same job as the previous
function but its a class that inherits from 'generic.DetailView' for a
display of a intermediate understanding of Django.
This type of class based views are not needed to complete the assignment
this is just an extra step I wanted to take

DELETED FUNCTION: purchase_detail
This function was replaced by the aformentioned class.
The creation of the class deemed this function useless so it is not
needed