urls.py/views.py change
urls.py:
Path for the home was changed because the class name was changed
Path for the dog tag list was changed to the use the new class
from views.py

views.py:
NEW CLASS: DogTagListView
This class will be replacing the previous function that was in its place
DogTagListView will do the exact same job as the previous
function but its a class that inherits from 'generic.ListView' for a
display of a intermediate understanding of Django.
This type of class based views are not needed to complete the assignment
this is just an extra step I wanted to take

DELETED FUNCTION: dog_tag_list
This function was replaced by the aformentioned class.
The creation of the class deemed this function useless so it is not
needed